# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Freight Audit Automation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transporters:
    get:
      operationId: listTransporters
      tags: [transporters]
      summary: List transporters
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: clientId
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of transporters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transporter"
    post:
      operationId: createTransporter
      tags: [transporters]
      summary: Create a transporter
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransporterBody"
      responses:
        "201":
          description: Created transporter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transporter"
  /transporters/{id}:
    patch:
      operationId: updateTransporter
      tags: [transporters]
      summary: Update a transporter
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransporterBody"
      responses:
        "200":
          description: Updated transporter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transporter"
    delete:
      operationId: deleteTransporter
      tags: [transporters]
      summary: Delete a transporter
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /rate-cards:
    get:
      operationId: listRateCards
      tags: [ratecards]
      summary: List rate cards
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: clientId
          schema:
            type: ["integer", "null"]
        - in: query
          name: transporterId
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of rate cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RateCard"
    post:
      operationId: createRateCard
      tags: [ratecards]
      summary: Create a rate card
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRateCardBody"
      responses:
        "201":
          description: Created rate card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateCard"
  /rate-cards/{id}:
    patch:
      operationId: updateRateCard
      tags: [ratecards]
      summary: Update a rate card
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRateCardBody"
      responses:
        "200":
          description: Updated rate card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateCard"
    delete:
      operationId: deleteRateCard
      tags: [ratecards]
      summary: Delete a rate card
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: clientId
          schema:
            type: ["integer", "null"]
        - in: query
          name: transporterId
          schema:
            type: ["integer", "null"]
        - in: query
          name: status
          schema:
            type: ["string", "null"]
        - in: query
          name: search
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice with BOL mappings and run audit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created invoice with audit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get a single invoice with full audit details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice with audit details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice (remark, resolved status)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /invoices/{id}/reaudit:
    post:
      operationId: reauditInvoice
      tags: [audit]
      summary: Re-run audit for an invoice
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Transporter:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        clientId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - clientId
        - createdAt
    CreateTransporterBody:
      type: object
      properties:
        name:
          type: string
        clientId:
          type: ["integer", "null"]
      required:
        - name
    UpdateTransporterBody:
      type: object
      properties:
        name:
          type: string
        clientId:
          type: ["integer", "null"]
    RateCard:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        transporterId:
          type: integer
        pricingType:
          type: string
          enum: [weight, distance, hybrid]
        ratePerKg:
          type: ["number", "null"]
        ratePerKm:
          type: ["number", "null"]
        fuelPerKm:
          type: ["number", "null"]
        effectiveFrom:
          type: ["string", "null"]
          format: date
        effectiveTo:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - transporterId
        - pricingType
        - ratePerKg
        - ratePerKm
        - fuelPerKm
        - effectiveFrom
        - effectiveTo
        - createdAt
    CreateRateCardBody:
      type: object
      properties:
        clientId:
          type: integer
        transporterId:
          type: integer
        pricingType:
          type: string
          enum: [weight, distance, hybrid]
        ratePerKg:
          type: ["number", "null"]
        ratePerKm:
          type: ["number", "null"]
        fuelPerKm:
          type: ["number", "null"]
        effectiveFrom:
          type: ["string", "null"]
        effectiveTo:
          type: ["string", "null"]
      required:
        - clientId
        - transporterId
        - pricingType
    UpdateRateCardBody:
      type: object
      properties:
        pricingType:
          type: string
          enum: [weight, distance, hybrid]
        ratePerKg:
          type: ["number", "null"]
        ratePerKm:
          type: ["number", "null"]
        fuelPerKm:
          type: ["number", "null"]
        effectiveFrom:
          type: ["string", "null"]
        effectiveTo:
          type: ["string", "null"]
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        clientId:
          type: integer
        transporterId:
          type: integer
        invoiceAmount:
          type: number
        calculatedAmount:
          type: ["number", "null"]
        difference:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
          enum: [PASS, WARNING, FAIL, null]
        remark:
          type: ["string", "null"]
        resolved:
          type: boolean
        invoiceDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        clientName:
          type: ["string", "null"]
        transporterName:
          type: ["string", "null"]
        bolCount:
          type: ["integer", "null"]
      required:
        - id
        - invoiceNumber
        - clientId
        - transporterId
        - invoiceAmount
        - calculatedAmount
        - difference
        - status
        - remark
        - resolved
        - invoiceDate
        - createdAt
        - clientName
        - transporterName
        - bolCount
    InvoiceDetail:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        clientId:
          type: integer
        transporterId:
          type: integer
        invoiceAmount:
          type: number
        calculatedAmount:
          type: ["number", "null"]
        difference:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
          enum: [PASS, WARNING, FAIL, null]
        remark:
          type: ["string", "null"]
        resolved:
          type: boolean
        invoiceDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        clientName:
          type: ["string", "null"]
        transporterName:
          type: ["string", "null"]
        bolCount:
          type: ["integer", "null"]
        bols:
          type: array
          items:
            $ref: "#/components/schemas/AuditedBol"
      required:
        - id
        - invoiceNumber
        - clientId
        - transporterId
        - invoiceAmount
        - calculatedAmount
        - difference
        - status
        - remark
        - resolved
        - invoiceDate
        - createdAt
        - clientName
        - transporterName
        - bolCount
        - bols
    AuditedBol:
      type: object
      properties:
        id:
          type: integer
        bolNumber:
          type: string
        origin:
          type: string
        destination:
          type: string
        distance:
          type: number
        actualWeight:
          type: number
        volumetricWeight:
          type: ["number", "null"]
        chargeableWeight:
          type: ["number", "null"]
        baseCost:
          type: ["number", "null"]
        fuelCost:
          type: ["number", "null"]
        calculatedCost:
          type: ["number", "null"]
        validationIssues:
          type: array
          items:
            type: string
      required:
        - id
        - bolNumber
        - origin
        - destination
        - distance
        - actualWeight
        - volumetricWeight
        - chargeableWeight
        - baseCost
        - fuelCost
        - calculatedCost
        - validationIssues
    CreateInvoiceBody:
      type: object
      properties:
        invoiceNumber:
          type: string
        clientId:
          type: integer
        transporterId:
          type: integer
        invoiceAmount:
          type: number
        invoiceDate:
          type: string
        bolNumbers:
          type: array
          items:
            type: string
      required:
        - invoiceNumber
        - clientId
        - transporterId
        - invoiceAmount
        - invoiceDate
        - bolNumbers
    UpdateInvoiceBody:
      type: object
      properties:
        remark:
          type: ["string", "null"]
        resolved:
          type: boolean
...