# lib/api-zod/src/generated/types/createInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Freight Audit Automation System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInvoiceBody {
  invoiceNumber: string;
  clientId: number;
  transporterId: number;
  invoiceAmount: number;
  invoiceDate: string;
  bolNumbers: string[];
}