# lib/api-zod/src/generated/types/invoiceDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Freight Audit Automation System API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditedBol } from "./auditedBol";
import type { InvoiceDetailStatus } from "./invoiceDetailStatus";

export interface InvoiceDetail {
  id: number;
  invoiceNumber: string;
  clientId: number;
  transporterId: number;
  invoiceAmount: number;
  /** @nullable */
  calculatedAmount: number | null;
  /** @nullable */
  difference: number | null;
  /** @nullable */
  status: InvoiceDetailStatus;
  /** @nullable */
  remark: string | null;
  resolved: boolean;
  invoiceDate: Date;
  createdAt: Date;
  /** @nullable */
  clientName: string | null;
  /** @nullable */
  transporterName: string | null;
  /** @nullable */
  bolCount: number | null;
  bols: AuditedBol[];
}